backend/monolith/requirements.txt
# FastAPI and server
fastapi==0.119.0
uvicorn[standard]==0.37.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
alembic==1.14.0

# Data validation
pydantic[email]==2.12.2

# Environment variables
python-dotenv==1.1.1

# Redis (optional for caching)
redis==5.2.1

# Email
aiosmtplib==3.0.2
mailgun2==1.2.1

# SMS (Twilio)
twilio==9.0.0

# HTTP client (for external APIs)
httpx==0.28.1
requests==2.31.0

# Security
bcrypt==4.1.2
passlib==1.7.4
python-jose[cryptography]==3.3.0

# Utilities
pyyaml==6.0.3

# Rate limiting
slowapi==0.1.9

# HTML Sanitization
bleach==6.1.0
ruff==0.1.15
Jinja2==3.1.4

# CSRF Protection
fastapi-csrf-protect==1.0.7
pydantic-settings==2.0.3

# Testing
pytest==8.0.0